+++ examples/cxx/publish_subscribe_dynamic_data/src/publisher.cpp
@@ -48,9 +48,9 @@ auto main() -> int {
 
     while (node.wait(CYCLE_TIME).has_value()) {
         const auto required_memory_size = (counter + 1) * (counter + 1); // NOLINT
-        auto sample = publisher.loan_slice_uninit(required_memory_size).expect("acquire sample");
+        auto sample = publisher.loan_slice_uninit(static_cast<uint64_t>(required_memory_size)).expect("acquire sample");
         auto initialized_sample =
-            sample.write_from_fn([&](auto byte_idx) { return (byte_idx + counter) % 255; }); // NOLINT
+            sample.write_from_fn([&](auto byte_idx) { return (byte_idx + static_cast<uint64_t>(counter)) % 255; }); // NOLINT
 
         send(std::move(initialized_sample)).expect("send successful");
