+++ iceoryx2-cxx/tests/src/service_request_response_tests.cpp
@@ -795,7 +795,7 @@ TYPED_TEST(ServiceRequestResponseTest, write_from_fn_works) {
     auto received_request = std::move(active_request.value());
     auto iterations = 0;
     for (const auto& item : received_request.payload()) {
-        ASSERT_THAT(item.a, Eq(DummyData::DEFAULT_VALUE_A + iterations));
+        ASSERT_THAT(item.a, Eq(DummyData::DEFAULT_VALUE_A + static_cast<uint64_t>(iterations)));
         ASSERT_THAT(item.z, Eq(iterations % 2 == 0));
         ++iterations;
     }
